+++ PrimeFactorizer/PrimeCheck.h
@@ -9,7 +9,7 @@ class PrimeCheck
 public:
 	PrimeCheck() noexcept; 
 	PrimeCheck(shared_ptr <PrimeLibrary> l);
-	bool is_prime(uint_fast64_t input);
+	bool is_prime(unsigned long long input);
 
 private:
 	shared_ptr <PrimeLibrary> Lib;
