+++ PrimeFactorizer/PrimeFactor.h
@@ -11,13 +11,13 @@ public:
 	PrimeFactor() ;//no lib
 	PrimeFactor(shared_ptr <PrimeLibrary> l);//0-4; select library size with 0 being no library and 4 being the largest
 	~PrimeFactor();
-	vector <uint_fast64_t> factor(uint_fast64_t num);
+	vector <unsigned long long> factor(unsigned long long num);
 
 private:
 
 	unique_ptr <PrimeCheck> Check;
 	unique_ptr <PrimeDivisor> Div;
-	vector <uint_fast64_t> Divisors;
+	vector <unsigned long long> Divisors;
 
 };
